Restructure OTLP exporter directory (#171)
