+++ exporters/otlp/src/recordable.cc
@@ -1,4 +1,4 @@
-#include "exporters/otlp/recordable.h"
+#include "opentelemetry/exporters/otlp/recordable.h"
 
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace exporter
