+++ exporters/otlp/test/otlp_exporter_test.cc
@@ -1,4 +1,4 @@
-#include "otlp_exporter.h"
+#include "opentelemetry/exporters/otlp/otlp_exporter.h"
 #include "opentelemetry/proto/collector/trace/v1/trace_service_mock.grpc.pb.h"
 #include "opentelemetry/sdk/trace/simple_processor.h"
 #include "opentelemetry/sdk/trace/tracer_provider.h"
